refactor(ComponentLibrary): Remove fragments directory
