+++ packages/react-component-library/src/components/Masthead/UserLink.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Avatar, AVATAR_VARIANT, Link } from '../../components'
+import { Avatar, AVATAR_VARIANT, Link } from ".."
 
 interface UserLinkProps {
   className?: string
