+++ packages/react-component-library/src/components/Masthead/index.tsx
