+++ packages/react-component-library/src/components/NotificationPanel/Notification.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, RenderResult } from '@testing-library/react'
 
-import { Link } from '../../components'
+import { Link } from ".."
 import { Notification } from '.'
 
 const NOW = '2019-11-05T11:00:00.000Z'
