+++ packages/react-component-library/src/components/Sidebar/Sidebar.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import '@testing-library/jest-dom/extend-expect'
 import { fireEvent, render, RenderResult } from '@testing-library/react'
 
-import { Link } from '../../components/Link'
+import { Link } from "../Link"
 import { Sidebar, SidebarNav, SidebarNavItem, SidebarUser } from '.'
 import { Graph, House, Tools } from '../../icons'
 import { Notification, Notifications } from '../NotificationPanel'
