--- packages/react-component-library/src/fragments/index.ts
@@ -1,31 +1,0 @@
-import {
-  Masthead,
-  MastheadNav,
-  MastheadNavItem,
-  MastheadUser,
-} from './Masthead'
-import {
-  Notification,
-  NotificationPanel,
-  Notifications,
-} from './NotificationPanel'
-import {
-  Sidebar,
-  SidebarNav,
-  SidebarNavItem,
-  SidebarUser,
-} from './Sidebar'
-
-export {
-  Masthead,
-  MastheadNav,
-  MastheadNavItem,
-  MastheadUser,
-  Notification,
-  NotificationPanel,
-  Notifications,
-  Sidebar,
-  SidebarNav,
-  SidebarNavItem,
-  SidebarUser,
-}
